Add MenuItem component
