+++ screens/IncidentScreen.js
@@ -1,13 +1,11 @@
 import React from 'react';
 import {
-  StyleSheet, Text, TouchableOpacity, View,
+  StyleSheet,
+  Text,
+  View,
 } from 'react-native';
-import { inject, observer } from 'mobx-react';
-import { Icon } from 'react-native-elements';
-import Icons from '@expo/vector-icons';
+import MenuItem from '../components/MenuItem';
 
-@inject('user', 'chat')
-@observer
 class IncidentScreen extends React.Component {
   static navigationOptions = {
     title: 'Incident',
@@ -20,7 +18,6 @@ class IncidentScreen extends React.Component {
   };
 
   onPressContact = () => {
-    console.log('onPressContact');
     const { navigation } = this.props;
     navigation.navigate('Contact');
   };
@@ -29,48 +26,16 @@ class IncidentScreen extends React.Component {
     console.log('call fleet manager');
   };
 
-  Item = ({
-    icon,
-    iconStyle,
-    label,
-    separator = true,
-  }) => {
-    const Icon = Icons[icon.split('/')[0]];
-    const iconName = icon.split('/')[1];
-
-    return (
-      <View style={[s.item, separator ? s.itemSeparator : {}]}>
-        <View style={{ flex: 1 }}>
-          <Icon
-            name={iconName}
-            style={[s.itemIcon, iconStyle]}
-          />
-        </View>
-        <Text style={s.itemLabel}>
-          {label}
-        </Text>
-      </View>
-    );
-  }
-
-  MenuItem = ({ onPress, ...rest }) => (
-    <TouchableOpacity
-      style={s.menuItemTouchable}
-      onPress={onPress}
-    >
-      <this.Item {...rest} />
-    </TouchableOpacity>
-  );
-
-
   render() {
-    const { MenuItem } = this;
-
     return (
       <View style={s.container}>
         <View style={s.innerContainer}>
-          <Text style={[s.messageText, { marginTop: 90 }]}>Emergency!</Text>
-          <Text style={[s.messageText, { marginTop: 10 }]}>The car will stop.</Text>
+          <Text style={[s.messageText, { marginTop: 90 }]}>
+            Emergency!
+          </Text>
+          <Text style={[s.messageText, { marginTop: 10 }]}>
+            The car will stop.
+          </Text>
           <Text style={[s.messageText, { marginTop: 10 }]}>
             Please contact your fleet manager via chat or phone call.
           </Text>
@@ -80,12 +45,12 @@ class IncidentScreen extends React.Component {
             label="Contact Fleet Manager"
             icon="Entypo/chat"
             onPress={this.onPressContact}
-            separator={false}
-          />
+            />
           <MenuItem
             label="Call Fleet Manager"
             icon="FontAwesome/phone"
             onPress={this.onPressCall}
+            separator={false}
           />
         </View>
       </View>
@@ -110,44 +75,8 @@ const s = StyleSheet.create({
     color: '#fefefe',
     textAlign: 'center',
   },
-  bottomButtons: {
-    position: 'absolute',
-    bottom: 60,
-    left: 20,
-    right: 20,
-    height: 75,
-    justifyContent: 'center',
-    alignItems: 'center',
-    flexDirection: 'row',
-  },
-  item: {
-    flexDirection: 'row',
-    justifyContent: 'center',
-    paddingBottom: 20,
-  },
-  itemSeparator: {
-    borderBottomWidth: 1,
-    borderBottomColor: '#C2C2C2',
-    paddingBottom: 15,
-    marginBottom: 15,
-  },
-  itemIcon: {
-    marginRight: 14,
-    color: '#ffffff',
-    fontSize: 26,
-  },
-  itemLabel: {
-    flex: 7,
-    alignSelf: 'center',
-    fontSize: 16,
-    fontWeight: '600',
-    color: '#ffffff',
-  },
-  menuItemTouchable: {
-    alignItems: 'center',
-  },
   menu: {
-    marginBottom: 20,
+    marginBottom: 40,
     marginHorizontal: 20,
     marginTop: 30,
   },
